+++ long_term_scheduler.cpp
@@ -1,7 +1,7 @@
 //============================================================================
-// Name				 : long_term_scheduler.cpp
-// Author			 :
-// Version		 :
+// Name		   : long_term_scheduler.cpp
+// Author	   :
+// Version	   :
 // Description : Source file for long term scheduler
 //============================================================================
 
@@ -15,29 +15,6 @@
 #define MEM_SIZE 16
 int main_memory[MEM_SIZE];
 int available_memory;
-
-/*
- * Primary vectors of PCBs for the NEW and READY queue.
- */
-std::vector<PCB> new_queue;
-std::vector<PCB> ready_queue;
-
-/*
- * Blocked queue for processes waiting on IO
- */
-std::vector<PCB> blocked_queue;
-
-/*
- * List of processes that have completed. May not be needed.
- * But used anyway to store the history of processes in the OS.
- */
-std::vector<PCB> finished_list;
-
-/*
- * The main running process.
- * Possibly a better way to implement this.
- */
-PCB running_process;
 
 /*
  * Counts the number of cycles for an IO Request
@@ -265,58 +242,6 @@ void check_interrupts () {
 				 io_cycle_count = 25;
 				 if (DEBUG) cout << "DEBUG: (check_interupts): Resetting io_cycle_count to " << io_cycle_count << endl;
 			}
-		}
-	}
-}
-
-/*
- * Basically FCFS temporarily, also not fully working.
- * This should make use of debug statements, since getting a good implementation
- * working that we also know is working will take some time.
- */
-void short_term_scheduler () {
-
-	// Check if there is no process currently running
-	if (running_process.check_state("NULL")) {
-
-		if (DEBUG) cout << "DEBUG: (short_term_scheduler): no process currently running" << endl;
-
-		// Check if the ready queue is empty
-		if (ready_queue.empty()) {
-
-			if (DEBUG) cout << "DEBUG: (short_term_scheduler): Ready Queue is currently empty" << endl;
-
-			// Check if the new queue is empty
-			if (new_queue.empty()) {
-
-				if (DEBUG) cout << "DEBUG: (short_term_scheduler): New Queue is also currently empty" << endl;
-
-				// Do nothing? Long term scheduler will eventually put a process into the new/ready queue?
-			}
-
-			// Nothing to run, short term scheduler did nothing.
-			return;
-		}
-
-		// Get the next process from the ready queue
-		running_process = ready_queue.front();
-
-		// Remove that process from the front of the ready queue
-		ready_queue.erase(ready_queue.begin());
-
-		if (DEBUG) cout << "DEBUG: (short_term_scheduler): put process: " << running_process.get_id() << " into running" << endl;
-	}
-	else {
-
-		// A process is already running
-		if (DEBUG) cout << "DEBUG: (short_term_scheduler): Current running process ID is :" << running_process.get_id() << endl;
-
-		// Check if that process has finished yet
-		if (running_process.runtime_remaining() == 0) {
-			if (DEBUG) cout << "DEBUG: (short_term_scheduler): Running Process has finished " << endl;
-
-			// The finished process now goes from READY->EXIT
-			process_exit();
 		}
 	}
 }
