+++ short_term_scheduler.h
@@ -1,0 +1,15 @@
+//============================================================================
+// Name        : short_term_scheduler.h
+// Author      :
+// Version     : 0.01
+// Description : Short term scheduling bringing processes READY to RUNNING
+//============================================================================
+
+#ifndef SHORT_TERM_SCHEDULER_H_
+#define SHORT_TERM_SCHEDULER_H_
+
+#include "state_queues.h"
+
+void short_term_scheduler();
+
+#endif /* SHORT_TERM_SCHEDULER_H_ */
